refactor: merge reset password
